requirements.txt
streamlit
pandas
psycopg2-binary
plotly
kafka-python
python-dotenv
